+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 - Exceptions are emailed. The recipients and subject prefix can be configured.
 - Inline editing for grids.
 - Tables can have a column named `extended_columns` with JSONB key/values that can be displayed and edited.
+- Added toggle camera scan route to RMD menu.
 ### Changed
+- Created RMD utilities route and moved check_barcode to it.
 ### Fixed
 
 ## [0.6.0] - 2019-02-01